+++ src/dlc/UI/InGame/CivVAccess_UnitControlCore.lua
@@ -46,6 +46,7 @@ UnitControl.preflightAttack = UnitControlCombat.preflightAttack
 UnitControl.preflightAttackTarget = UnitControlCombat.preflightAttackTarget
 UnitControl.enemyAt = UnitControlMovement.enemyAt
 UnitControl.enemyCityAt = UnitControlMovement.enemyCityAt
+UnitControl.actionKeyLabel = UnitControlMovement.actionKeyLabel
 UnitControl.registerPending = UnitControlMovement.registerPending
 UnitControl.notifyDeferredCommit = UnitControlMovement.notifyDeferredCommit
 UnitControl.notifyCommitCanceled = UnitControlMovement.notifyCommitCanceled